lms/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-tinymce==4.1.0
pillow==11.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
